requirements.txt
# Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Config
python-dotenv==1.0.1

# OpenAI
openai==1.51.0
httpx==0.27.2

# PDF parsing
pymupdf==1.24.10

# PPTX parsing
python-pptx==1.0.2

# DOCX parsing
python-docx==1.1.2

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Data validation
pydantic==2.9.2